# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
#host: mysoccer-api.appspot.com
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /hello:
    # binds a127 app logic to a route
    x-swagger-router-controller: hello_world
    get:
      description: Returns 'Hello' to the caller
      # used as the method name of the controller
      operationId: hello
      parameters:
        - name: name
          in: query
          description: The name of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    
  /soccer:
    x-swagger-router-controller: soccer.controller
    get:
      description: get the game system list
      operationId: getSoccers
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetSoccerListResponse"
        default:
          description: Error
          schema: 
            $ref: "#/definitions/ErrorResponse" 
    post:
      description: add a new game system to the list
      operationId: addSoccer
      parameters:
        - in: body
          name: soccer
          description: The Soccer to create.
          schema:
            $ref: "#/definitions/Soccer"
      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/GetSoccerResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /soccer/{id}:
    x-swagger-router-controller: soccer.controller
    get:
      description: get a game system
      operationId: getSoccerbyId
      parameters:
        - name: id
          type: number
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetSoccerResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: update a soccer
      operationId: updateSoccer
      # define the parameters
      parameters:
        - name: id
          description: soccer id
          type: number
          in: path
          required: true
        - in: body
          name: soccer
          description: The Soccer to update.
          schema:
            $ref: "#/definitions/Soccer"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetSoccerResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    delete:
      description: delete a soccer
      operationId: deleteSoccer
      # define the parameters
      parameters:
        - name: id
          description: soccer id
          type: number
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  Soccer:
    type: object
    properties:
      name:
        type: string
        description: Name of the Soccer
      
      team:
        type: string
        description: Description of the Soccer
      league:
        type: string
        description: Description of the Soccer
      dtechnical:
        type: string 
        description: Description oh the Soccer 
      createdAt:
        type: string 
        description: Description oh the Soccer 
      updatedAt:
        type: string 
        description: Description oh the Soccer 
        
  GetSoccerResponse:
    type: object
    properties:
      id:
        type: number
        description: identifier
      name:
        type: string
        description: Name of the Soccer
      team:
        type: string
        description: Description of the Soccer
      league:
        type: string
        description: Description of the Soccer
      dtechnical:
        type: string 
        description: Description oh the Soccer 

  
      
  GetSoccerListResponse:
    required:
      - soccer
    properties:
      soccer:
        type: array
        items:
          $ref: "#/definitions/GetSoccerResponse"
  GeneralResponse:
    type: object
    properties:
      success:
        type: number
        description: returns 1 if successful
      description:
        type: string
        description: a short comment
    required:
      - success
      - description



  HelloWorldResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string